requirements.txt
# Core dependencies
phidata==2.4.22
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
typer==0.12.3
rich==13.7.1

# LLM providers
google-cloud-aiplatform==1.57.0
openai==1.35.7
anthropic==0.30.0

# Additional tools
tavily-python==0.3.3

# Networking and API
requests==2.32.3
httpx==0.27.0

# Data handling
numpy==2.0.0

# File operations
python-multipart==0.0.9

# Async support
anyio==4.4.0

# CLI enhancements
click==8.1.7
shellingham==1.5.4

# Time handling
python-dateutil==2.9.0.post0

# pgvector 
# pypdf 
# psycopg2-binary 
# sqlalchemy 
# fastapi